Refactor: 기술, 직무 도메인을 career 하위에서 동일 선상으로 변경
